Исправил рандомную генерацию картинок в форме точки маршрута
