+++ src/components/utils.js
@@ -8,8 +8,12 @@ const getRandomIntegerNumber = (min, max) => {
   return min + Math.floor(Math.random() * (max - min));
 };
 
-const getRandomNumber = (length) => {
+const getRandomNumberFloor = (length) => {
+  return Math.floor(Math.random() * length);
+};
+
+const getRandomNumberCeil = (length) => {
   return Math.ceil(Math.random() * length);
 };
 
-export {getRandomArrayItem, getRandomNumber};
+export {getRandomArrayItem, getRandomNumberFloor, getRandomNumberCeil};
