+++ src/main.js
@@ -4,7 +4,7 @@ import {createSiteFilterTemplate} from "./components/site-filter.js";
 import {createSiteSortTemplate} from "./components/site-sort.js";
 import {createSiteFormTemplate} from "./components/site-form.js";
 import {createPointTemplate} from "./components/point.js";
-import {generatePoint} from "./mock/point.js";
+import {generatePoint} from "./mock/point-form.js";
 
 const point = generatePoint();
